Implemented blocking FIFO as a channel similiar to blocking interface
