+++ example/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(Framer)
 add_subdirectory(SingleMasterMultiSlave)
 add_subdirectory(FPI)
 add_subdirectory(FIFO)
+add_subdirectory(FIFO_Shared)
 add_subdirectory(tests)
 add_subdirectory(Whishbone)
 add_subdirectory(I2C)
