+++ example/FIFO/Old/Producer.h
@@ -1,9 +1,9 @@
-#include "Interfaces.h"
-#include "FIFO_Channel.h"
+#include "../Interfaces_new/Interfaces.h"
+
 #include "systemc.h"
 
 SC_MODULE(Producer){
-    sc_port<FIFO_out_if<int>> out;
+  blocking_out<int> out;
 
   int cnt = 1;
 
@@ -12,7 +12,7 @@ SC_MODULE(Producer){
           //Produce Values and write them to the Input of the FIFO
           out->write(cnt++);
           std::cout << "At " << sc_time_stamp() << " Producer sent: " << cnt-1 << endl;
-          wait(WAIT_TIME, SC_NS);
+          wait(1, SC_NS);
       }
   }
