+++ install/install_new.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #BINARY LOCATIONS
 SCAM_HOME="$HOME/DeSCAM" #NO SLASH AT THE END
-CMAKE_BIN="cmake" #NO SLASH AT THE END
+CMAKE_BIN="cmake3" #NO SLASH AT THE END
 PYTHON3="python3" #NO SLASH AT THE END
 
 #UNZIPPING FILES
